// src/types/modalType.ts
export enum ModalTypes {
  LOGIN = 'login',
  REGISTER = 'register',
}

export enum SortType {
  Popularity = 'popularity',
  VoteCount = 'vote_count',
  Title = 'original_title',
  ReleaseDate = 'release_date',
  Runtime = 'runtime',
  VoteAverage = 'average_rating',
}

export enum SortOrder {
  Asc = 1,
  Desc = -1,
}